+++ lib/pages/settings/language_settings_tile.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:gcrdeviceconfigurator/data/database.dart';
-import 'package:gcrdeviceconfigurator/data/language_settings.dart';
+import 'package:gcrdeviceconfigurator/data/settings.dart';
 import 'package:gcrdeviceconfigurator/i18n/languages.dart';
 
 class LanguageSettingTile extends StatelessWidget {
